// src/types/dataverse.ts
export interface PsDeliveryroutes {
  ps_deliveryroutesid?: string
  createdby?: string | null
  createdbyname?: string | null
  createdbyyominame?: string | null
  createdon?: string | null
  createdonbehalfby?: string | null
  createdonbehalfbyname?: string | null
  createdonbehalfbyyominame?: string | null
  importsequencenumber?: number | null
  modifiedby?: string | null
  modifiedbyname?: string | null
  modifiedbyyominame?: string | null
  modifiedon?: string | null
  modifiedonbehalfby?: string | null
  modifiedonbehalfbyname?: string | null
  modifiedonbehalfbyyominame?: string | null
  overriddencreatedon?: string | null
  ownerid?: string | null
  owneridname?: string | null
  owneridtype?: string | null
  owneridyominame?: string | null
  owningbusinessunit?: string | null
  owningbusinessunitname?: string | null
  owningteam?: string | null
  owninguser?: string | null
  ps_sequence?: number | null
  ps_sitelat?: number | null
  ps_sitelong?: number | null
  ps_vehiclename?: string | null
  statecode?: number | null
  statecodename?: unknown
  statuscode?: number | null
  statuscodename?: unknown
  timezoneruleversionnumber?: number | null
  utcconversiontimezonecode?: number | null
  versionnumber?: number | null
}

export type PsDeliveryroutesEntitySet = 'ps_deliveryrouteses'

export type PsDeliveryroutesCreate = Omit<PsDeliveryroutes, 'ps_deliveryroutesid' | 'createdon' | 'modifiedon' | 'versionnumber' | 'createdby' | 'modifiedby' | 'ownerid'>
// ...
export const psdeliveryroutesEntitySet: PsDeliveryroutesEntitySet = 'ps_deliveryrouteses'
// ...
export const EntitySets = {
  psdeliveryroutes: 'ps_deliveryrouteses' as const,
  psdeliveryschedule: 'ps_deliveryschedules' as const,
  psjobassetattachment: 'ps_jobassetattachments' as const,
  salesorder: 'salesorders' as const,
  salesorderdetail: 'salesorderdetails' as const,
  pspostalareaboundary: 'ps_postalareaboundaries' as const,
  psrelatedasset: 'ps_relatedassets' as const,
  pssetting: 'ps_settings' as const,
  psterritorygroup: 'ps_territorygroups' as const,
  psvehicledatabase: 'ps_vehicledatabases' as const,
} as const